+++ frontend/src/views/sql-editor/EditorPanel/GutterBar/GutterBar.vue
