+++ frontend/src/views/sql-editor/EditorPanel/GutterBar/common.ts
