+++ frontend/src/views/sql-editor/EditorPanel/GutterBar/index.ts
@@ -1,0 +1,3 @@
+import GutterBar from "./GutterBar.vue";
+
+export default GutterBar;
