+++ frontend/src/views/sql-editor/EditorPanel/types.ts
