--- frontend/src/views/sql-editor/EditorPanelBar/index.ts
@@ -1,3 +1,0 @@
-import EditorPanelBar from "./EditorPanelBar.vue";
-
-export default EditorPanelBar;
